[IDLE-278] feat: 릴리즈 업데이트 기능 적용

+++ app-ui/lib/ui/setting/setting_screen.dart
@@ -8,6 +8,7 @@ import 'package:mybrary/ui/common/layout/default_layout.dart';
 import 'package:mybrary/ui/profile/profile_edit/profile_edit_screen.dart';
 import 'package:mybrary/ui/setting/components/account_withdrawal.dart';
 import 'package:mybrary/ui/setting/components/login_info.dart';
+import 'package:mybrary/utils/logics/common_utils.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class SettingScreen extends StatelessWidget {
@@ -88,7 +89,7 @@ class SettingScreen extends StatelessWidget {
                     _settingTab(
                       tabTitle: '1:1 문의하기',
                       onTap: () async {
-                        await _connectWebLink(
+                        await connectWebLink(
                           webLink: inquiryLink,
                         );
                       },
@@ -133,7 +134,7 @@ class SettingScreen extends StatelessWidget {
                       tabTitle: '리뷰로 응원하기',
                       onTap: () async {
                         // Todo: 추후 AOS, IOS 별 링크 설정 필요
-                        await _connectWebLink(
+                        await connectWebLink(
                           webLink: androidAppLink,
                         );
                       },
@@ -178,19 +179,6 @@ class SettingScreen extends StatelessWidget {
     );
   }
 
-  Future<void> _connectWebLink({
-    required String webLink,
-  }) async {
-    String url = webLink;
-    if (await canLaunchUrl(Uri.parse(url))) {
-      await launchUrl(
-        Uri.parse(url),
-        mode: LaunchMode.externalApplication,
-        webOnlyWindowName: '_self',
-      );
-    }
-  }
-
   Future<dynamic> _showLogoutAlert(
       BuildContext context, FlutterSecureStorage secureStorage) async {
     return await showDialog(
@@ -216,14 +204,14 @@ class SettingScreen extends StatelessWidget {
           actions: [
             Row(
               children: [
-                _confirmButton(
+                confirmButton(
                   onTap: () {
                     Navigator.of(context).pop();
                   },
                   buttonText: '취소',
                   isCancel: true,
                 ),
-                _confirmButton(
+                confirmButton(
                   onTap: () async {
                     await secureStorage.deleteAll();
 
@@ -236,6 +224,8 @@ class SettingScreen extends StatelessWidget {
                   },
                   buttonText: '로그아웃',
                   isCancel: false,
+                  confirmButtonColor: commonRedColor,
+                  confirmButtonText: commonWhiteColor,
                 ),
               ],
             ),
@@ -268,37 +258,6 @@ class SettingScreen extends StatelessWidget {
               'assets/svg/icon/profile_menu_arrow.svg',
             ),
           ],
-        ),
-      ),
-    );
-  }
-
-  Widget _confirmButton({
-    required GestureTapCallback? onTap,
-    required String buttonText,
-    required bool isCancel,
-  }) {
-    return Expanded(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 4.0),
-        child: InkWell(
-          onTap: onTap,
-          child: Container(
-            height: 46.0,
-            decoration: BoxDecoration(
-              color: isCancel ? greyF1F2F5 : commonRedColor,
-              borderRadius: BorderRadius.circular(4.0),
-            ),
-            child: Center(
-              child: Text(
-                buttonText,
-                style: commonSubBoldStyle.copyWith(
-                  color: isCancel ? commonBlackColor : commonWhiteColor,
-                  fontSize: 14.0,
-                ),
-              ),
-            ),
-          ),
         ),
       ),
     );
